WIP fix unit test
